assorted recommended changes
