--- my-app/src/Components/Button.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const Button = (props) => {
-  return (
-    <button onClick={props.handleClick}>{props.buttonText}</button>
-  );
-}
-
-export default Button;
