+++ package.json
@@ -2,6 +2,7 @@
   "name": "my-app",
   "version": "0.1.0",
   "private": true,
+  "proxy": "https://cat-fact.herokuapp.com",
   "dependencies": {
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
