+++ src/Components/Button.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export const Button = ({handleClick, buttonText = ''}) => {
+  return (
+    <button onClick={handleClick}>{buttonText}</button>
+  );
+};
