+++ src/Components/index.js
@@ -1,0 +1,3 @@
+export * from './Button';
+export * from './Card';
+export * from './Toggle';
